+++ src/main/io/osd_dji_hd.c
@@ -497,6 +497,8 @@ static char * osdArmingDisabledReasonMessage(void)
         //     return OSD_MESSAGE_STR("HARDWARE FAILURE");
         case ARMING_DISABLED_BOXFAILSAFE:
             return OSD_MESSAGE_STR("FAILSAFE ENABLED");
+        case ARMING_DISABLED_DRONECAN_BRIDGE:
+            return OSD_MESSAGE_STR("CAN BRIDGE ACTIVE");
         case ARMING_DISABLED_RC_LINK:
             return OSD_MESSAGE_STR("NO RC LINK");
         case ARMING_DISABLED_THROTTLE: